# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Api Documentation",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Laravel Api"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "admin@admin.com"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "admin"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/logout": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User logout",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/configuration": {
            "post": {
                "tags": [
                    "Configuration"
                ],
                "summary": "Fees configuration",
                "operationId": "configuration",
                "parameters": [
                    {
                        "name": "boleto_fee",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1.45"
                    },
                    {
                        "name": "credit_card_fee",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "7.63"
                    },
                    {
                        "name": "fee_amount_less",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "2"
                    },
                    {
                        "name": "fee_amount_less_value",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "3000"
                    },
                    {
                        "name": "fee_amount_greater",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    },
                    {
                        "name": "fee_amount_greater_value",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "3000"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/exchange/available": {
            "get": {
                "tags": [
                    "Currency Exchange"
                ],
                "summary": "Check the currencies available to exchange",
                "operationId": "available",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/exchange": {
            "post": {
                "tags": [
                    "Currency Exchange"
                ],
                "summary": "Exchange the currencies",
                "operationId": "exchange",
                "parameters": [
                    {
                        "name": "income_currency",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "USD"
                    },
                    {
                        "name": "amount_exchange",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "5000"
                    },
                    {
                        "name": "payment_method",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "boleto"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/exchange/history": {
            "get": {
                "tags": [
                    "Currency Exchange"
                ],
                "summary": "Check the exchange history",
                "operationId": "history",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "You need a token to access the API",
                "name": "bearer",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}